Add files via upload
these are the html code calling javascript for the interactive dashboard. These files are calling the KSI_final.csv and KSI_geojson files in the code. I worked on the files by calling them from the folder as webserver. The terminal commands are "npm install -g http-server" and "http-server". everytime the files are being executed, the Cache memory in the browser has to be cleared out.

Interactive bar chart shows the count of records in years based on districts.
Interactive heatmap shows the count of the records based on the years selected.
Interactive bubble map shows the bubble based on the year selected.